♻️ extract noopRecorderApi and mockCiVisibilityWindowValues
