+++ packages/rum-core/src/domain/rumEventsCollection/view/viewCollection.spec.ts
@@ -1,10 +1,11 @@
 import type { Duration, RelativeTime, ServerDuration, TimeStamp } from '@datadog/browser-core'
 import type { RecorderApi } from '../../../boot/rumPublicApi'
 import type { TestSetupBuilder } from '../../../../test/specHelper'
-import { noopRecorderApi, setup } from '../../../../test/specHelper'
+import { setup } from '../../../../test/specHelper'
 import type { RawRumViewEvent } from '../../../rawRumEvent.types'
 import { RumEventType, ViewLoadingType } from '../../../rawRumEvent.types'
 import { LifeCycleEventType } from '../../lifeCycle'
+import { noopRecorderApi } from '../../../../test/noopRecorderApi'
 import type { ViewEvent } from './trackViews'
 import { startViewCollection } from './viewCollection'
