+++ packages/rum-core/test/noopRecorderApi.ts
@@ -1,0 +1,10 @@
+import type { RecorderApi } from '@datadog/browser-rum-core'
+import { noop } from '@datadog/browser-core'
+
+export const noopRecorderApi: RecorderApi = {
+  start: noop,
+  stop: noop,
+  isRecording: () => false,
+  onRumStart: noop,
+  getReplayStats: () => undefined,
+}
